backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.0
pydantic>=2.7.3
pydantic-settings>=2.3.1
python-dotenv>=1.0.1
google-generativeai>=0.5.4
alembic>=1.13.1
httpx>=0.27.0
requests>=2.31.0
